aggregator: Rework flush-stop logic
Process it in two steps, first in pre_queue where it takes the srcpad stream
lock, resets internal state on the sinkpad, and queues up the event if it's been
received on all sinkpads, and then again in the task (in its previous place)
where it resets the aggregator and gets forwarded downstream.

Solve race where the sinkpad gets hit with a flush-stop while aggregate is
running.

Co-authored-by: Vivia Nikolaidou <vivia.nikolaidou@ltnglobal.com>
Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/612>