citemplate: port fedora build jobs to a parallel matrix
This is a new feature which makes it so we can generate
jobs based on the possible matrix of the environment variables
we pass into it.

In this commit we refactored the gstbuild template to a matrix of
Buildtype, debugbuild (and could have also set werror, but we
always have it on in fedora gstbuilds).

Then create 2 jobs, one for each compiler set. We could have
put them in the matrix, but CC and CXX are kinda coupled
and doesn't make sense to tests the matrix between them.

https://docs.gitlab.com/ce/ci/yaml/README.html#parallel-matrix-jobs

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/4281>