Revert "v4l2object: Only offer inactive pools and if needed"
This reverts commit 85b9893e963ea5d342289e318daeefba35ff4a07.

This change caused a regression reported in #796. Basically the pool get
deactivated by basesrc and the encoder does not recover. Recovering with
current design would cause insertion of an unwanted keyframe.

This is being reverted in 1.18 only, a proper solution is wanted for 1.20.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-good/-/merge_requests/785>