avauddec: Check planar-ness of frame rather than context
The check seems to be to present to verify that the decoded frame
matches the format we expect. The actual check for the layout of the
frame was being performed against the context instead.

The check fails at least for avdec_aptx_hd, where the AVCodecContext has
the sample format set to AV_SAMPLE_FMT_NONE.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-libav/-/merge_requests/107>